Adjust code after PR
