+++ Sig.App.Backend/Requests/Queries/Cards/CardBalanceReport.cs
@@ -1,0 +1,10 @@
+﻿using Sig.App.Backend.DbModel.Entities.Cards;
+
+namespace Sig.App.Backend.Requests.Queries.Cards
+{
+    public class CardBalanceReport
+    {
+        public Card Card { get; set; }
+        public decimal Total { get; set; }
+    }
+}
